Support version prefix for Mandrel
This enables setting version to `23` or `23.0` and setup-graalvm will
find the latest mandrel version on that train,
e.g. `mandrel-23.1.0.0-Final` for `23` and `mandrel-23.0.1.2-Final` for
`23.0`.

This feature is useful for CI jobs testing an application/framework
working with an older Mandrel LTS version while still using the latest
CPU update of it.